backend/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Ambro Quach",email = "ambroquach@berkeley.edu"}
]
package-mode = false
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "joblib (>=1.5.2,<2.0.0)",
    "fastapi-simple-redis-cache (>=1.3.3,<2.0.0)",
    "torch (>=2.9.0,<3.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "supabase (>=2.0.0,<3.0.0)",
]

[tool.poetry.group.dev.dependencies]
datasets = "^4.3"
scikit-learn = "^1.7.2"
pytest = "^8.4"
ruff = "^0.14"
pre-commit = "^4.3"
evaluate = "^0.4"
accelerate = "^1.11.0"
python-dotenv = "^1.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
